refactor: P2 server split 5/N — extract workspace route domain
Move /api/workspace/* (tree/file/raw/download-folder/office-preview/
pptx-preview + mutations), /api/workspaces CRUD, and the
/api/sessions/:id/workspace binding routes out of server.ts into
server/routes/workspaces.ts as handleWorkspacesRoutes. The registry and
session-binding blocks move up ~400 lines to the single delegation point;
all matches are exact-path, so the reordering is inert.

Domain-private helpers move with it (workspaceIdFromQuery/Body,
workspaceRelativePath, workspaceSkillNode, readWorkspaceTree, zipDirectory,
the PPTX→SVG converter family). safeWorkspacePath becomes a per-request
closure over the injected workspaceRegistry so route bodies stay verbatim.
contentDispositionAttachment and WORKSPACE_IGNORES are shared with
unextracted domains, so they land in server/file-helpers.ts.

server.ts: 3985 → 3276 lines. Smoke coverage: /api/workspaces list,
workspace tree on the tmp workspace, missing-file 404, and the 400 guard
on deleting the shared tmp workspace.
